manager/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
psutil==5.9.8
PyYAML==6.0.2
httpx==0.27.0
starlette==0.37.2
python-multipart==0.0.9
rich==13.7.1
colorama==0.4.6
websockets==12.0 